test: split mobile terminal lifecycle coverage
